Update namespace for move and fix dispose warnings
