+++ Mapsui.Rendering.Skia/Caching/TileCacheEntry.cs
@@ -1,4 +1,4 @@
-﻿using Mapsui.Rendering.Skia.Cache;
+﻿using Mapsui.Rendering.Caching;
 using SkiaSharp;
 
 namespace Mapsui.Rendering.Skia.Caching;
@@ -6,20 +6,7 @@ namespace Mapsui.Rendering.Skia.Caching;
 public sealed class TileCacheEntry(SKObject skObject) : ITileCacheEntry
 {
     private readonly SKObject _skObject = skObject;
-    private bool _disposed;
 
     public object Object => _skObject;
     public long IterationUsed { get; set; }
-
-    public void Dispose()
-    {
-        if (_disposed)
-        {
-            return;
-        }
-#pragma warning disable IDISP007
-        _skObject.Dispose();
-#pragma warning restore IDISP007
-        _disposed = true;
-    }
 }
