--- Mapsui.Rendering.Skia/GlobalSuppressions.cs
@@ -1,19 +1,0 @@
-﻿// This file is used by Code Analysis to maintain SuppressMessage
-// attributes that are applied to this project.
-// Project-level suppressions either have no target or are given
-// a specific target and scoped to a namespace, type, member, etc.
-
-using System.Diagnostics.CodeAnalysis;
-
-[assembly: SuppressMessage("IDisposableAnalyzers.Correctness", "IDISP001:Dispose created.", Justification = "<Pending>", Scope = "member", Target = "~M:Mapsui.Rendering.Skia.CalloutStyleRenderer.UpdateContent(Mapsui.Styles.CalloutStyle)")]
-[assembly: SuppressMessage("IDisposableAnalyzers.Correctness", "IDISP001:Dispose created.", Justification = "<Pending>", Scope = "member", Target = "~M:Mapsui.Rendering.Skia.BitmapRenderer.Draw(SkiaSharp.SKCanvas,SkiaSharp.SKImage,SkiaSharp.SKRect,System.Single)")]
-[assembly: SuppressMessage("IDisposableAnalyzers.Correctness", "IDISP015:Member should not return created and cached instance.", Justification = "<Pending>", Scope = "member", Target = "~M:Mapsui.Rendering.Skia.BitmapRenderer.GetPaint(System.Single,System.Boolean@)~SkiaSharp.SKPaint")]
-[assembly: SuppressMessage("IDisposableAnalyzers.Correctness", "IDISP004:Don't ignore created IDisposable.", Justification = "<Pending>", Scope = "member", Target = "~M:Mapsui.Utilities.SvgHelper.LoadSvg(System.String)~Svg.Skia.SKSvg")]
-[assembly: SuppressMessage("IDisposableAnalyzers.Correctness", "IDISP003:Dispose previous before re-assigning", Justification = "<Pending>", Scope = "member", Target = "~M:Mapsui.Rendering.Skia.SkiaWidgets.ScaleBarWidgetRenderer.Draw(SkiaSharp.SKCanvas,Mapsui.Viewport,Mapsui.Widgets.IWidget,System.Single)")]
-[assembly: SuppressMessage("IDisposableAnalyzers.Correctness", "IDISP003:Dispose previous before re-assigning", Justification = "<Pending>", Scope = "member", Target = "~M:Mapsui.Rendering.Skia.MapRenderer.#ctor(Mapsui.Rendering.IRenderCache)")]
-[assembly: SuppressMessage("IDisposableAnalyzers.Correctness", "IDISP004:Don't ignore created IDisposable", Justification = "<Pending>", Scope = "member", Target = "~M:Mapsui.Rendering.Skia.MapRenderer.#ctor")]
-[assembly: SuppressMessage("IDisposableAnalyzers.Correctness", "IDISP001:Dispose created", Justification = "<Pending>", Scope = "member", Target = "~M:Mapsui.Rendering.Skia.CalloutStyleRenderer.RenderCallout(Mapsui.Styles.CalloutStyle,Mapsui.Rendering.DrawableImageCache)")]
-[assembly: SuppressMessage("IDisposableAnalyzers.Correctness", "IDISP001:Dispose created", Justification = "<Pending>", Scope = "member", Target = "~M:Mapsui.Rendering.Skia.PolygonRenderer.Draw(SkiaSharp.SKCanvas,Mapsui.Viewport,Mapsui.Styles.VectorStyle,Mapsui.IFeature,NetTopologySuite.Geometries.Polygon,System.Single,Mapsui.Rendering.IVectorCache)")]
-[assembly: SuppressMessage("IDisposableAnalyzers.Correctness", "IDISP001:Dispose created", Justification = "<Pending>", Scope = "member", Target = "~M:Mapsui.Rendering.Skia.PolygonRenderer.GetImage(Mapsui.Rendering.DrawableImageCache,System.Int32)~SkiaSharp.SKImage")]
-[assembly: SuppressMessage("IDisposableAnalyzers.Correctness", "IDISP001:Dispose created", Justification = "<Pending>", Scope = "member", Target = "~M:Mapsui.Rendering.Skia.SymbolStyleRenderer.DrawImage(SkiaSharp.SKCanvas,Mapsui.Viewport,Mapsui.Layers.ILayer,System.Double,System.Double,Mapsui.Styles.SymbolStyle,Mapsui.Rendering.DrawableImageCache)~System.Boolean")]
-[assembly: SuppressMessage("IDisposableAnalyzers.Correctness", "IDISP001:Dispose created", Justification = "<Pending>", Scope = "member", Target = "~M:Mapsui.Rendering.Skia.SymbolStyleRenderer.DrawImage(SkiaSharp.SKCanvas,Mapsui.Viewport,Mapsui.Layers.ILayer,System.Double,System.Double,Mapsui.Styles.SymbolStyle,Mapsui.Rendering.DrawableImageCache)~System.Boolean")]
